rename trusted client api to userstore
